small fix update doc script

+++ scripts/update_doc.sh
@@ -22,7 +22,10 @@ mkdir -p ${TARGET_DIR}/html
 rm -rf ${TARGET_DIR}/html/*
 
 # clear old logs
-rm log*
+files=(log*)
+if [[ -f "${files[0]}" ]]; then
+    rm log*
+fi
 
 MODULES=(Main Cpp Py C Csharp)
 # generate tag files
